Found a way to stop the game if there is a result.
I had to create a DOM Element inside of gameLogic as a "global" variable to use it in resultCheck and playTurn, then, i constructed an algorithm, if there is "x" element inside of the textContent of the DOM Element,
remove the Event Listeners of the grid cells, after trying many ways to solve this, i ended up making a variable called endGame inside of resultCheck, if there is a winner, this variable is 1, if there is a draw, this variable is 2.
so when a game is over (endGame != 0), we change the DOM Element to "1" (obviously this is not the final content).

Now, inside of playTurn, i had to convert the arrow function inside of the EventListener that contains the turn logic, into a separated function inside of playTurn, all of this in order to being able of removing this EventListener when the DOM Element = "1"
I had to do the EventListener to remove the other EventListeners inside of the forEach method so i'll be able to reach the "box" parameter.
then i made a loop through all of the boxClick elements, to remove their EventListeners one by one.